requirements.txt
gitpython
ipython
matplotlib>=3.2.2
numpy>=1.18.5
opencv-python>=4.1.1
psutil
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
thop>=0.1.1
tqdm>=4.64.0
tensorboard>=2.4.1
pandas>=1.1.4
seaborn>=0.11.0
timm ==0.9.16
pytorch-lightning >=2.0.0 
nltk >=3.7.0  
PyYAML >=6.0.0 
pillow==10.2.0
tqdm==4.66.2
torch==2.2.1+cpu
torchmetrics==1.3.1
torchvision==0.17.1+cpu